src/key_information_extractor/PICK-pytorch/requirements.txt
torchvision==0.6.1
tabulate==0.8.7
overrides==3.0.0
allennlp==1.0.0
torchtext==0.6.0
